[M68K] Renaming aim to have consistent naming convetion
Fixup compilation as name are case senistive for most arch ...
