+++ arch/M68K/M68Kdasm.h
@@ -38,7 +38,7 @@
 /* ======================================================================== */
 
 /* Import the configuration for this build */
-#include "m68kconf.h"
+#include "M68Kconf.h"
 
 
 /* ======================================================================== */
